Making job creation more robust to ensure search results remain valid.
We use square brackets as part of the search functionality to signal that there is a hit with a search term. We therefor want to avoid having names in our DB which contain these special characters.

We could validate this on the backend as well to be double sure but as a first pass I believe on mobile should be good enough.